htdocs/www.kitabnow.com/requirements.txt
Flask==2.1.3
Flask-Login==0.6.3
Flask-SQLAlchemy-Meiqia==2016.8.1
SQLAlchemy==1.3.23
Werkzeug==2.0.3
gunicorn==19.7.1